Passing in main span ref via rpc server info and rename tracer package
